scoring/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.1.1
psycopg2-binary==2.9.9
shapely==2.0.6
numpy==1.26.4
httpx==0.27.2
pydantic==2.9.2
pygeohash==1.1.0
pytest==8.3.3
pytest-asyncio==0.24.0
fakeredis==2.26.1
